django/pyproject.toml
[tool.poetry]
name = "reactdjangofullstack"
version = "0.1.0"
description = ""
authors = ["Xaron11 <kaczmarzykdawid4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
